+++ AmazonChimeSDK/AmazonChimeSDK/audiovideo/DefaultAudioVideoController.swift
@@ -38,16 +38,21 @@ import Foundation
     public func start() throws {
         // By default, start for calls without CallKit integration. Use start(callKitEnabled:)
         // to override the default behavior if the call is integrated with CallKit
-        try self.start(callKitEnabled: false)
+        try self.start(audioVideoConfiguration: AudioVideoConfiguration())
     }
 
     public func start(callKitEnabled: Bool) throws {
+        try self.start(audioVideoConfiguration: AudioVideoConfiguration(callKitEnabled: callKitEnabled))
+    }
+
+    public func start(audioVideoConfiguration: AudioVideoConfiguration) throws {
         try audioClientController.start(audioFallbackUrl: configuration.urls.audioFallbackUrl,
                                         audioHostUrl: configuration.urls.audioHostUrl,
                                         meetingId: configuration.meetingId,
                                         attendeeId: configuration.credentials.attendeeId,
                                         joinToken: configuration.credentials.joinToken,
-                                        callKitEnabled: callKitEnabled)
+                                        callKitEnabled: audioVideoConfiguration.callKitEnabled,
+                                        audioMode: audioVideoConfiguration.audioMode)
         videoClientController.subscribeToVideoTileControllerObservers(observer: videoTileController)
         videoClientController.start()
     }