+++ AmazonChimeSDK/AmazonChimeSDK/internal/audio/DefaultAudioClientObserver.swift
@@ -171,6 +171,7 @@ class DefaultAudioClientObserver: NSObject, AudioClientDelegate {
 
         if let attendeesWithJoinedStatus = newAttendeeMap[AttendeeStatus.joined] {
             let attendeesJoined = attendeesWithJoinedStatus.subtracting(currentAttendeeSet)
+            logger.info(msg: "attendeesJoined: \(attendeesJoined)")
             if !attendeesJoined.isEmpty {
                 ObserverUtils.forEach(observers: realtimeObservers) { (observer: RealtimeObserver) in
                     observer.attendeesDidJoin(attendeeInfo: [AttendeeInfo](attendeesJoined))
@@ -181,6 +182,7 @@ class DefaultAudioClientObserver: NSObject, AudioClientDelegate {
 
         if let attendeesWithLeftStatus = newAttendeeMap[AttendeeStatus.left] {
             if !attendeesWithLeftStatus.isEmpty {
+                logger.info(msg: "attendeesLeft: \(attendeesWithLeftStatus)")
                 ObserverUtils.forEach(observers: realtimeObservers) { (observer: RealtimeObserver) in
                     observer.attendeesDidLeave(attendeeInfo: [AttendeeInfo](attendeesWithLeftStatus))
                 }
@@ -190,6 +192,7 @@ class DefaultAudioClientObserver: NSObject, AudioClientDelegate {
 
         if let attendeesWithDroppedStatus = newAttendeeMap[AttendeeStatus.dropped] {
             if !attendeesWithDroppedStatus.isEmpty {
+                logger.info(msg: "attendeesDropped: \(attendeesWithDroppedStatus)")
                 ObserverUtils.forEach(observers: realtimeObservers) { (observer: RealtimeObserver) in
                     observer.attendeesDidDrop(attendeeInfo: [AttendeeInfo](attendeesWithDroppedStatus))
                 }