+++ AmazonChimeSDK/AmazonChimeSDKTests/internal/contentshare/DefaultContentShareVideoClientControllerTests.swift
@@ -39,7 +39,8 @@ class DefaultContentShareVideoClientControllerTests: CommonTestCase {
                                      token: self.joinToken,
                                      sending: false,
                                      config: any(),
-                                     appInfo: any())).wasCalled()
+                                     appInfo: any(),
+                                     signalingUrl: any())).wasCalled()
         verify(videoClientMock.setExternalVideoSource(any())).wasCalled()
         verify(videoClientMock.setSending(true)).wasCalled()
     }
@@ -49,7 +50,8 @@ class DefaultContentShareVideoClientControllerTests: CommonTestCase {
                                     token: any(),
                                     sending: any(),
                                     config: any(),
-                                    appInfo: any())).will {_, _, _, _, _ in
+                                    appInfo: any(),
+                                    signalingUrl: any())).will {_, _, _, _, _, _ in
             self.defaultContentShareVideoClientController.videoClientDidConnect(nil, controlStatus: 1)
         }
         given(videoClientMock.stop()).will {
@@ -63,7 +65,8 @@ class DefaultContentShareVideoClientControllerTests: CommonTestCase {
                                      token: self.joinToken,
                                      sending: false,
                                      config: any(),
-                                     appInfo: any())).wasCalled()
+                                     appInfo: any(),
+                                     signalingUrl: any())).wasCalled()
         verify(videoClientMock.setExternalVideoSource(any())).wasCalled(2)
         verify(videoClientMock.setSending(true)).wasCalled(2)
     }
@@ -73,7 +76,8 @@ class DefaultContentShareVideoClientControllerTests: CommonTestCase {
                                     token: any(),
                                     sending: any(),
                                     config: any(),
-                                    appInfo: any())).will {_, _, _, _, _ in
+                                    appInfo: any(),
+                                    signalingUrl: any())).will {_, _, _, _, _, _ in
             self.defaultContentShareVideoClientController.videoClientDidConnect(nil, controlStatus: 1)
         }