+++ AmazonChimeSDKDemo/AmazonChimeSDKDemo/JoiningViewController.swift
@@ -15,10 +15,13 @@ class JoiningViewController: UIViewController, UITextFieldDelegate {
     @IBOutlet var meetingIdTextField: UITextField!
     @IBOutlet var nameTextField: UITextField!
     @IBOutlet var versionLabel: UILabel!
-    @IBOutlet var joinWithoutCallKitButton: UIButton!
-    @IBOutlet var joinAsIncomingCallButton: UIButton!
-    @IBOutlet var joinAsOutgoingCallButton: UIButton!
+    @IBOutlet var callKitOptionsPicker: UIPickerView!
+    @IBOutlet var audioModeOptionsPicker: UIPickerView!
+    @IBOutlet var joinButton: UIButton!
     @IBOutlet var debugSettingsButton: UIButton!
+
+    var callKitOptions = ["Don't use CallKit", "CallKit as Incoming in 10s", "CallKit as Outgoing"]
+    var audioModeOptions = ["Stereo/48KHz Audio", "Mono/48KHz Audio", "Mono/16KHz Audio"]
 
     private let toastDisplayDuration = 2.0
     private let mainStoryboard = UIStoryboard(name: "Main", bundle: nil)
@@ -29,26 +32,35 @@ class JoiningViewController: UIViewController, UITextFieldDelegate {
         meetingIdTextField.delegate = self
         nameTextField.delegate = self
 
+        callKitOptionsPicker.delegate = self
+        callKitOptionsPicker.dataSource = self
+
+        audioModeOptionsPicker.delegate = self
+        audioModeOptionsPicker.dataSource = self
+
         setupHideKeyboardOnTap()
         versionLabel.text = "amazon-chime-sdk-ios@\(Versioning.sdkVersion())"
     }
 
     override func viewWillAppear(_: Bool) {
-        setJoinButtons(isEnabled: true)
+        callKitOptionsPicker.selectRow(0, inComponent: 0, animated: false)
+        joinButton.isEnabled = true
     }
 
-    @IBAction func joinWithoutCallKitButtonClicked(_: UIButton) {
-        joinMeeting(callKitOption: .disabled)
-    }
+    @IBAction func joinButton(_: UIButton) {
+        // CallKit Option
+        let callKitOption = getSelectedCallKitOption()
+        if (callKitOption == .incoming) {
+            view.makeToast("You can background the app or lock screen while waiting",
+                           duration: incomingCallKitDelayInSeconds)
+        }
 
-    @IBAction func joinAsIncomingCallButton(_: UIButton) {
-        view.makeToast("You can background the app or lock screen while waiting",
-                       duration: incomingCallKitDelayInSeconds)
-        joinMeeting(callKitOption: .incoming)
-    }
+        // Audio Mode
+        let audioMode = getSelectedAudioMode()
 
-    @IBAction func joinAsOutgoingCallButton(_: UIButton) {
-        joinMeeting(callKitOption: .outgoing)
+        joinMeeting(audioVideoConfig: AudioVideoConfiguration(audioMode: audioMode, callKitEnabled: callKitOption != .disabled),
+                    callKitOption: callKitOption
+        )
     }
 
     @IBAction func debugSettingsButtonClicked (_: UIButton) {
@@ -66,13 +78,29 @@ class JoiningViewController: UIViewController, UITextFieldDelegate {
         return true
     }
 
-    private func setJoinButtons(isEnabled: Bool) {
-        joinWithoutCallKitButton.isEnabled = isEnabled
-        joinAsIncomingCallButton.isEnabled = isEnabled
-        joinAsOutgoingCallButton.isEnabled = isEnabled
+    func getSelectedCallKitOption() -> CallKitOption {
+        switch callKitOptionsPicker.selectedRow(inComponent: 0) {
+        case 1:
+            return .incoming
+        case 2:
+            return .outgoing
+        default:
+            return .disabled
+        }
     }
 
-    func joinMeeting(callKitOption: CallKitOption) {
+    func getSelectedAudioMode() -> AudioMode {
+        switch audioModeOptionsPicker.selectedRow(inComponent: 0) {
+        case 1:
+            return .mono48K
+        case 2:
+            return .mono16K
+        default:
+            return .stereo48K
+        }
+    }
+
+    func joinMeeting(audioVideoConfig: AudioVideoConfiguration, callKitOption: CallKitOption) {
         view.endEditing(true)
         let meetingId = meetingIdTextField.text ?? ""
         let name = nameTextField.text ?? ""
@@ -87,6 +115,7 @@ class JoiningViewController: UIViewController, UITextFieldDelegate {
 
         MeetingModule.shared().prepareMeeting(meetingId: meetingId,
                                               selfName: name,
+                                              audioVideoConfig: audioVideoConfig,
                                               option: callKitOption,
                                               overriddenEndpoint: debugSettingsModel.endpointUrl) { success in
             DispatchQueue.main.async {
@@ -96,6 +125,40 @@ class JoiningViewController: UIViewController, UITextFieldDelegate {
                                         duration: self.toastDisplayDuration)
                 }
             }
+        }
+    }
+}
+
+extension JoiningViewController: UIPickerViewDelegate {
+    func pickerView(_ pickerView: UIPickerView, titleForRow row: Int, forComponent _: Int) -> String? {
+        if pickerView == callKitOptionsPicker {
+            if row >= callKitOptions.count {
+                return nil
+            }
+            return callKitOptions[row]
+        } else if pickerView == audioModeOptionsPicker {
+            if row >= audioModeOptions.count {
+                return nil
+            }
+            return audioModeOptions[row]
+        } else {
+            return nil
+        }
+    }
+}
+
+extension JoiningViewController: UIPickerViewDataSource {
+    func numberOfComponents(in _: UIPickerView) -> Int {
+        return 1
+    }
+
+    func pickerView(_ pickerView: UIPickerView, numberOfRowsInComponent _: Int) -> Int {
+        if pickerView == callKitOptionsPicker {
+            return callKitOptions.count
+        } else if pickerView == audioModeOptionsPicker {
+            return audioModeOptions.count
+        } else {
+            return 0
         }
     }
 }